file type icon fix
